This is the intial commit of an empty repository for the PrimeTest Project!
This project will build a simple python app that test a range of numbers,
starting at 1, to see if they are prime numbers.   The high end of the test
range is specifiable.  Upon completion, the program will print the detected
prime numbers to the screen.  An option to export the list of detected primes
to a file will also be included.

This is a very straigtforward python project suitable for python &
software development students of all kinds.   The real purpose of this
project is complementary to program being coded.

Management of development projects via source control tools like Git and
GitHub greatly simplify the management of collaborative development projects.
reduce the likelyhood of management errors that can cost hours of lost productivity,
make dealing with errors and conflicts much easier, and provide many, many, other
features, as well, that simplify project management while increasing productivity.
This project was created with the goal if providing a solid introduction to the
source control process, in general, and to Git & GitHub specifically.

-Jonathan Fierstein                       5/1/2024